requirements.txt
# Lecture Notes Pipeline - Python Dependencies
# Tested with Python 3.10+

# Speech Recognition
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# LLM Client
ollama>=0.3.0

# Audio Processing
ffmpeg-python>=0.2.0

# CLI & Progress
rich>=13.0.0
tqdm>=4.66.0
click>=8.1.0

# Configuration
pyyaml>=6.0.0

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0
